refactor: begin cleanup
