+++ src/instruments/src/EFB/Dashboard/Widgets/Reminders/WeatherReminder.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import { WeatherWidget } from '../WeatherWidget';
+import { RemindersSection } from './RemindersSection';
+import { useAppSelector } from '../../../Store/store';
+
+export const WeatherReminder = () => {
+    const { departingAirport, arrivingAirport } = useAppSelector((state) => state.simbrief.data);
+    const { userDepartureIcao, userDestinationIcao } = useAppSelector((state) => state.dashboard);
+
+    return (
+        <RemindersSection title="Weather" noLink>
+            <div className="space-y-6">
+                <WeatherWidget name="origin" simbriefIcao={departingAirport} userIcao={userDepartureIcao} />
+                <div className="w-full h-1 rounded-full bg-theme-accent" />
+                <WeatherWidget name="destination" simbriefIcao={arrivingAirport} userIcao={userDestinationIcao} />
+            </div>
+        </RemindersSection>
+    );
+};
